simplify source setup + more validation of scripts
If source.pretty_name isn't set it's assumed it's not actually a source
for whatever reason, along with this, if the proxy argument isn't set,
it'll inherit it from the settings file. It is like this, in case
someone wants to use spiderpirate like a node module, and they can
thereby have their own proxy settings system etc...

On top of this, for leetx and rarbg, we no longer assume that it's
settings proxy value is the same as what was searched with, so now the
search function passes it along instead, for the same reasons as before.