actually interpret sizes
Instead of just spitting back the sizes as the string that was picked
up, we now try to interpret it, and convert it to bytes, to the client
can show and format it however it wants universally without problems.

Overall this is better for the API

I also implemented this on the frontend of course.